Move queryBuilder mock to separate file and adjust it's references in codebase;
